+++ changelog.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ## 97bc4346faf5cfe7c3e87cdba0a31da248221d2d - 2017-11-28 - Yudhvir
 ### Added
+- Firebase integration
+- Static kits displaying
+- Clustering
+
+## 97bc4346faf5cfe7c3e87cdba0a31da248221d2d - 2017-11-28 - Yudhvir
+### Added
 - Map
 - Help me button
 - Location